+++ packages/plugins/EVM/src/state/NameService.ts
@@ -1,50 +1,9 @@
-import ENS from 'ethjs-ens'
 import type { Subscription } from 'use-subscription'
 import { getEnumAsArray } from '@dimensiondev/kit'
 import type { Plugin } from '@masknet/plugin-infra'
-import { NameServiceResolver, NameServiceState } from '@masknet/plugin-infra/web3'
-import { ChainId, formatEthereumAddress, isValidAddress, isZeroAddress, ProviderType } from '@masknet/web3-shared-evm'
-import { Providers } from './Connection/provider'
-
-class ENS_Resolver implements NameServiceResolver<ChainId> {
-    private ens: ENS | null = null
-
-    private async createENS() {
-        if (this.ens) return this.ens
-        const provider = await Providers[ProviderType.MaskWallet].createWeb3Provider({
-            chainId: ChainId.Mainnet,
-        })
-        this.ens = new ENS({
-            provider,
-            network: ChainId.Mainnet,
-        })
-        return this.ens
-    }
-
-    async lookup(chainId: ChainId, name: string) {
-        if (chainId !== ChainId.Mainnet) return
-
-        try {
-            const ens = await this.createENS()
-            return ens.lookup(name)
-        } catch {
-            return
-        }
-    }
-
-    async reverse(chainId: ChainId, address: string) {
-        if (chainId !== ChainId.Mainnet) return
-
-        try {
-            const ens = await this.createENS()
-            const name = await ens.reverse(address)
-            if (!name.endsWith('.eth')) return
-            return name
-        } catch {
-            return
-        }
-    }
-}
+import { NameServiceState } from '@masknet/plugin-infra/web3'
+import { ChainId, formatEthereumAddress, isValidAddress, isZeroAddress } from '@masknet/web3-shared-evm'
+import { ENS_Resolver } from './NameService/ENS'
 
 export class NameService extends NameServiceState<ChainId> {
     constructor(
