+++ packages/plugins/Solana/src/state/NameService.ts
@@ -1,14 +1,9 @@
+import type { Subscription } from 'use-subscription'
 import type { Plugin } from '@masknet/plugin-infra'
-import { NameServiceResolver, NameServiceState } from '@masknet/plugin-infra/web3'
+import { NameServiceState } from '@masknet/plugin-infra/web3'
 import { ChainId, formatAddress, isValidAddress, isValidDomain, isZeroAddress } from '@masknet/web3-shared-solana'
 import { getEnumAsArray } from '@dimensiondev/kit'
-import type { Subscription } from 'use-subscription'
-import { lookup, reverse } from '../apis'
-
-export class BonfidaResolver implements NameServiceResolver<ChainId> {
-    lookup = lookup
-    reverse = reverse
-}
+import { BonfidaResolver } from './NameService/Bonfida'
 
 export class NameService extends NameServiceState<ChainId> {
     constructor(
